remove declarations and fix lint
