--- src/component.d.ts
@@ -1,3 +1,0 @@
-import { Props } from './types';
-
-export default function ReactGlobe(props: Props): JSX.Element;
