--- src/hooks.d.ts
@@ -1,6 +1,0 @@
-import { Size } from './types';
-
-export function useResize(
-  mountRef: React.RefObject<HTMLDivElement>,
-  initialSize?: Size,
-): Size;
