+++ src/index.js
@@ -4,5 +4,4 @@ export { default as Globe } from './globe';
 export { default as Tooltip } from './tooltip';
 
 export * from './defaults';
-export * from './types';
 export * from './utils';
